Moved Camera logic to a special handler object
This is somewhat big rewrite, but I felt like it
was necessary... the main Activity was becoming too
large and I didn't like it.  With this new design,
the activity doesn't even import
android.hardware.Camera

All of this is done towards the goal of being able
to register to a front facing camera.